// types/LiquidityPool.ts
export type WriteOptionEvent = TypedEvent<
  [string, BigNumber, BigNumber, BigNumber, string],
  {
    series: string;
    amount: BigNumber;
    premium: BigNumber;
    escrow: BigNumber;
    buyer: string;
  }
>;